+++ src/common/helpers/profile.helper.ts
@@ -1,30 +1,24 @@
 import { ResourceType } from '@common/constants/record.constants';
-import { PROFILES_CONFIG } from '@src/configs';
+import { PROFILE_CONFIG } from '@src/configs';
 
 export const getProfileConfig = ({
-  profileName,
   resourceType,
   profileId,
   referenceProfileId,
 }: {
-  profileName: keyof typeof PROFILES_CONFIG;
   resourceType?: ResourceType;
   profileId?: string | null;
   referenceProfileId?: string;
 }) => {
-  if (!PROFILES_CONFIG[profileName]) {
-    throw new Error(`Profile with ID ${profileName} does not exist in the configuration.`);
-  }
-
-  const { api, rootEntry } = PROFILES_CONFIG[profileName];
+  const { defaultProfileIds, rootEntry } = PROFILE_CONFIG;
   let ids: number[] = [];
 
   if (resourceType === ResourceType.work) {
-    ids = typeof profileId === 'string' ? [Number(profileId)] : [api.work];
+    ids = typeof profileId === 'string' ? [Number(profileId)] : [defaultProfileIds.work];
   } else if (resourceType === ResourceType.instance) {
-    ids = typeof profileId === 'string' ? [Number(profileId)] : [api.instance];
+    ids = typeof profileId === 'string' ? [Number(profileId)] : [defaultProfileIds.instance];
   } else {
-    ids = [api.profile];
+    ids = [];
   }
 
   if (referenceProfileId) {